+++ src/Data/List/Ext.hs
@@ -1,5 +1,5 @@
 module Data.List.Ext ( imap
-                     , ifilter
+                     , ifilter'
                      , prior
                      ) where
 
@@ -9,5 +9,5 @@ prior op xs = zipWith op (tail xs) xs
 imap :: (Int -> a -> b) -> [a] -> [b]
 imap f xs = fmap (uncurry f) (zip [1..] xs)
 
-ifilter :: (Int -> a -> Bool) -> [a] -> [a]
-ifilter p xs = snd <$> filter (uncurry p) (zip [1..] xs)
+ifilter' :: (Int -> a -> Bool) -> [a] -> [(Int, a)]
+ifilter' p xs = filter (uncurry p) (zip [1..] xs)