+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Allow `fn...` declarations with no arguments and no parentheses
   * Add conditionals
   * Fix bug in normalizing `Some` and `None`
+  * Fix bug in indexing + filter
 
 # 0.3.0.0